requirements.txt
Django==2.2.9
django-heroku==0.3.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4

